+++ matrix.h
@@ -3,4 +3,4 @@ double findOpedelitel(double **matrix, int size);
 // Нахождения ранга матрицы matrix (sizeY x sizeX)
 int findRand(int **matrix, int sizeY, int sizeX);
 // Нахождение произведение двух матриц
-int** power(int **firstMatrix,int firstSizeY, int firstSizeX, int **secondMatrix,int secondSizeY, int secondSizeX);
+double** power(double **firstMatrix,int firstSizeY, int firstSizeX, double **secondMatrix,int secondSizeY, int secondSizeX);
